Refactoing: auth as separated module to be pluginable
