+++ ols/src/auth/__init__.py
@@ -1,0 +1,1 @@
+"""Various implementations of auth module."""
