+++ tests/unit/auth/__init__.py
@@ -1,0 +1,1 @@
+"""Unit tests for auth. dependency."""
